node-fetch: normalize pem key objects from agent options
Hoist normalizePemKeyOption from node:tls into internal/tls and run the
agent's key option through it so key: [{ pem, passphrase }] reaches the
native tls config as plain PEM entries. Read init.agent with an
own-property check like the rest of the agent extraction. Covers the
pem-object form in the mTLS test.
